+++ model/user.js
@@ -35,6 +35,4 @@ const userSchema = new mongoose.Schema({
 
 
 const User = new mongoose.model("User", userSchema);
-module.exports={
-    User
-}
+module.exports=User
